# src/backend/TestPlanService/Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:5000",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/projects/{projectId}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_DeleteProject",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestSuites"
        ],
        "operationId": "TestSuites_MoveSuite",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "moveSuiteRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MoveSuiteRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/test/configs": {
      "get": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_GetConfigs",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetConfigsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_AddConfig",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdateConfigsRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/configvars": {
      "get": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_GetConfigsVars",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetConfigsVarsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_AddConfigVariable",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdateConfigVarRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/configs/{configId}": {
      "put": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_UpdateConfig",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "configId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdateConfigsRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_DeleteConfig",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "configId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/configvars/{id}": {
      "put": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_UpdateConfigVariable",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdateConfigVarRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TestConfigs"
        ],
        "operationId": "TestConfigs_DeleteConfigVariable",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "GetConfigsResponse": {
      "type": "object",
      "properties": {
        "configs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConfigItem"
          }
        }
      }
    },
    "ConfigItem": {
      "type": "object",
      "required": [
        "id",
        "isDefault"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "isDefault": {
          "type": "boolean"
        },
        "variables": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConfigVariableItem"
          }
        }
      }
    },
    "ConfigVariableItem": {
      "type": "object",
      "required": [
        "id",
        "variableId",
        "valueId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "variableId": {
          "type": "integer",
          "format": "int32"
        },
        "valueId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "GetConfigsVarsResponse": {
      "type": "object",
      "properties": {
        "vars": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConfigVarItem"
          }
        }
      }
    },
    "ConfigVarItem": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConfigVarValueItem"
          }
        }
      }
    },
    "ConfigVarValueItem": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "AddOrUpdateConfigsRequest": {
      "type": "object",
      "required": [
        "isDefault"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "isDefault": {
          "type": "boolean"
        },
        "variables": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AddOrUpdateConfigVariable"
          }
        }
      }
    },
    "AddOrUpdateConfigVariable": {
      "type": "object",
      "required": [
        "variableId",
        "valueId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "variableId": {
          "type": "integer",
          "format": "int32"
        },
        "valueId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "AddOrUpdateConfigVarRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AddOrUpdateConfigVarValue"
          }
        }
      }
    },
    "AddOrUpdateConfigVarValue": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "value": {
          "type": "string"
        }
      }
    },
...
    "MoveSuiteRequest": {
      "type": "object",
      "required": [
        "newParentId"
      ],
      "properties": {
        "newParentId": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}